refactored removeSolutionFile

+++ src/ObjectHandler/SolutionHandler.ts
@@ -77,7 +77,7 @@ export class SolutionHandler implements ISPObjectHandler {
             SP.Publishing.DesignPackage.install(clientContext, clientContext.get_site(), packageInfo, filerelativeurl);
             clientContext.executeQueryAsync(
                 (sender, args) => {
-                    this.removeSolutionFile(clientContext, filerelativeurl)
+                    this.removeSolutionFile(solutionConfig, clientContext, filerelativeurl)
                         .then(() => { Util.Resolve<void>(resolve, solutionConfig.Title, `Activated Solution with title : '${solutionConfig.Title}'.`); })
                         .catch((error) => {
                             Util.Reject<void>(reject, solutionConfig.Title,
@@ -105,17 +105,24 @@ export class SolutionHandler implements ISPObjectHandler {
         });
     }
 
-    private removeSolutionFile(clientContext: SP.ClientContext, fileRelativeUrl: string) {
+    private removeSolutionFile(solutionConfig: ISolution, clientContext: SP.ClientContext, fileRelativeUrl: string) {
         return new Promise<boolean>((resolve, reject) => {
             let item = clientContext.get_web().getFileByServerRelativeUrl(fileRelativeUrl);
-            item.deleteObject();
-            clientContext.executeQueryAsync(
-                (sender, args) => {
-                    resolve();
-                },
-                (sender, args) => {
-                    reject();
-                });
+            if (!item.get_serverObjectIsNull) {
+                item.deleteObject();
+                clientContext.executeQueryAsync(
+                    (sender, args) => {
+                        resolve();
+                    },
+                    (sender, args) => {
+                        Util.Reject<void>(reject, solutionConfig.Title,
+                            `Error while deleting Solution with the title '${solutionConfig.Title}': ${args.get_message()} '\n' ${args.get_stackTrace()}`);
+                    });
+            } else {
+                Util.Reject<void>(reject, solutionConfig.Title,
+                    `Error while deleting Solutionfile '${solutionConfig.FileName}'' - file not found`);
+            }
+
         });
     }
 }
